youtube_digest/backend/requirements.txt
fastapi
uvicorn
openai
youtube-transcript-api
google-api-python-client
python-dotenv
requests
